pyproject.toml
[project]
name = "pricing-derogation-dashboard"
version = "0.1.0"
description = "A modular Dash dashboard for pricing derogation analytics."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dash>=2.14.0",
    "dash-bootstrap-components>=1.5.0",
    "pandas>=2.1.0",
    "plotly>=5.18.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.1.0",
    "cachetools>=5.3.2",
    "apscheduler>=3.10.4",
    "mlflow>=2.9.0",
    "dvc>=3.33.0",
    "requests>=2.31.0",
    "O365>=2.0.30",
    "ibm-cos-sdk>=2.13.3",
    "dominodatalab>=1.3.0",
    "pyyaml>=6.0.1",
    "openpyxl>=3.1.2",
    "gunicorn>=21.2.0",
    "waitress>=2.1.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
